+++ go.mod
@@ -17,9 +17,10 @@ require (
 	github.com/stretchr/testify v1.3.0
 	google.golang.org/grpc v1.21.1
 	gopkg.in/yaml.v1 v1.0.0-20140924161607-9f9df34309c0
-	k8s.io/api v0.0.0-20190620073856-dcce3486da33
+	k8s.io/api v0.0.0-20190802060718-d0d4f3afa3ab
 	k8s.io/apiextensions-apiserver v0.0.0-20190325193600-475668423e9f
-	k8s.io/apimachinery v0.0.0-20190620073744-d16981aedf33
-	k8s.io/client-go v0.0.0-20190620074045-585a16d2e773
+	k8s.io/apimachinery v0.0.0-20190802060556-6fa4771c83b3
+	k8s.io/client-go v0.0.0-20190802021151-fdb3fbe99e1d
 	k8s.io/klog v0.3.3
+	k8s.io/kubectl v0.0.0-20190803022817-1937123dfffc
 )
